Other: Extract `addListViewToDropdown()` to own file.
