+++ tests/dropdown/manual/dropdown.js
@@ -15,11 +15,9 @@ import alignRightIcon from '@ckeditor/ckeditor5-core/theme/icons/object-right.sv
 import alignCenterIcon from '@ckeditor/ckeditor5-core/theme/icons/object-center.svg';
 import ButtonView from '../../../src/button/buttonview';
 
-import {
-	addListViewToDropdown,
-	addToolbarToDropdown,
-} from '../../../src/dropdown/utils';
+import { addToolbarToDropdown } from '../../../src/dropdown/utils';
 
+import addListViewToDropdown from '../../../src/dropdown/helpers/addlistviewtodropdown';
 import createDropdownView from '../../../src/dropdown/helpers/createdropdownview';
 import createButtonForDropdown from '../../../src/dropdown/helpers/createbuttonfordropdown';
 import closeDropdownOnBlur from '../../../src/dropdown/helpers/closedropdownonblur';
